requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-dotenv
streamlit
requests
pandas
plotly
